Fix URL of Wotton House website
It looks like the previous URL redirects to DeVere, which does not
seem to run Wotton House anymore.

The new website does not have a good URL specifically for the
location as far as I can see, so use the same one in both places
(it's easy enough to find by scrolling).