Enable the registration form for 2026
Inspiration from commit 5b629e1 but with this year's
typeform ID.

I did not add the style attribute as it no longer helps avoid
scroll bars (i.e. we get scroll bars with or without it
unfortunately, when in previous years it made the embeded form
cleaner).

And the opacity likewise did not seem to have any effect so I left
it out.